refactor: split shortlink services and harden gateway event delivery
Replace the legacy project stack with dedicated command, redirect, Leaf
segment allocation, risk, and Kafka/Flink/ClickHouse analytics modules.
Align Admin and Agent tools with the shared authorization and statistics APIs.

Add bounded APISIX event batching, acknowledgement accounting, worker
profiles, deployment manifests, component tests, and performance tooling.
Retain the recorded E2E and capacity results, including the failed 5k gate.

Validation: existing unit/component and bounded E2E/performance records in
plan/production-refactor reports 04 through 13; staged git diff --check passes.
This commit is a development checkpoint, not production capacity acceptance.

+++ admin/src/main/java/com/jupiter/shortlink/admin/dto/resp/ShortLinkGroupRespDTO.java
@@ -2,28 +2,18 @@ package com.jupiter.shortlink.admin.dto.resp;
 
 import lombok.Data;
 
-/**
- * 短链接分组返回实体对象
- */
+/** 短链接分组返回实体对象 */
 @Data
 public class ShortLinkGroupRespDTO {
-    /**
-     * 短链接分组标识
-     */
+    /** 短链接分组标识 */
     private String gid;
 
-    /**
-     * 分组名称
-     */
+    /** 分组名称 */
     private String name;
 
-    /**
-     * 分组排序
-     */
+    /** 分组排序 */
     private Integer sortOrder;
 
-    /**
-     * 分组下短链接数量
-     */
-    private Integer shortLinkCount;
+    /** 分组下短链接数量 */
+    private Long shortLinkCount;
 }
